requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
distlib==0.3.6
Django==3.0.6
django-mssql-backend==2.8.1
djangorestframework==3.14.0
filelock==3.12.0
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
Pillow==9.5.0
platformdirs==3.5.1
PyJWT==2.7.0
pyodbc==4.0.39
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
twilio==8.2.2
urllib3==2.0.2
virtualenv==20.23.0
yarl==1.9.2
